Don't change default backend when it fails to load
If using the gconf backend and it fails to load libcompizconfig will automatically fall back to the ini backend. However, it will also update the default backend to the ini backend so the user will have to manually switch back once the problem is fixed. This patch makes the default backend not get updated to ini if the regular one failed to load. Thanks to Michael Vogt for the patch.

Backport of 7f990c801d3b1dbd99811223182be1067f8970f8 to compiz-0.8